+++ src/plugin_handler.js
@@ -1,4 +1,3 @@
-import path from 'path';
 export class Plugins {
   static get plugins() {
     if (!Plugins.prototype.__plugins) {
@@ -6,43 +5,12 @@ export class Plugins {
     }
     return Plugins.prototype.__plugins;
   }
-  static registerGlobalPlugin(registerablePlugin)  {
-    let plugin;
-    if (typeof registerablePlugin === 'object') {
-      plugin = registerablePlugin;
-    } else {
-      plugin = requirePluginFromPath(registerablePlugin);
-    }
+  static registerGlobalPlugin(plugin)  {
     if (validatePlugin(plugin)) {
       Plugins.plugins.push(plugin);
     } else {
-      console.warn('Plugin is not a valid Serves Up plugin:', registerablePlugin);
+      console.warn('Plugin is not a valid Serves Up plugin');
     }
-  }
-}
-
-function resolvePluginPath(pluginPath) {
-  if (path.isAbsolute(pluginPath) || pluginPath.includes('./')) {
-    return path.join(process.cwd(),pluginPath);
-  }
-  try {
-    return require.resolve(pluginPath);
-  } catch(e) {
-    console.error(e);
-    return null;
-  }
-}
-
-function requirePluginFromPath(pluginPath) {
-  try {
-    console.log(pluginPath);
-    const resolvedPath = resolvePluginPath(pluginPath);
-    const loadedPlugin = require(require.resolve(resolvedPath));
-    loadedPlugin.__plugin_name = pluginPath;
-    return loadedPlugin;
-  }catch(e) {
-    console.error(e);
-    return null;
   }
 }
